requirements.txt
alethic-ism-core==1.0.33
alethic-ism-db==1.0.34
python-dotenv
restrictedpython
psycopg2-binary
pyyaml
pydantic
nats-py
mako